【修改描述】按照客户要求修改入参结构体 【DTS/CR/AR单号】AR20240119615151 【合入版本】BeiMing-K 24.0.0
Match-id-daf70f9d212dffe31df7a1c455daa16e6ce19318

+++ src/sdk/mirror_client.h
@@ -53,7 +53,7 @@ public:
     explicit MirrorClient(int32_t type) : mDeployType(type) {}
     ~MirrorClient() = default;
 
-    uint16_t SelectingPt(uint32_t objectId, AffinityStrategy affinity);
+    uint16_t SelectingPt(uint64_t objectId, AffinityStrategy affinity);
 
     uint16_t ParseLocation(Bio::ObjLocation location);
